[DOP-3711]: Add primary unique index and add bulk upsert
